+++ apps/backoffice-v2/src/domains/ui-definition/hooks/queries/useTranslateUiDefinitionQuery/useTranslateUiDefinitionQuery.tsx
@@ -1,0 +1,16 @@
+import { useQuery } from '@tanstack/react-query';
+
+import { uiDefinitionQueryKeys } from '@/domains/ui-definition/query-keys';
+
+export const useTranslateUiDefinitionQuery = ({
+  id,
+  partialUiDefinition,
+}: {
+  id: string;
+  partialUiDefinition: Record<string, unknown>;
+}) => {
+  return useQuery({
+    ...uiDefinitionQueryKeys.translate({ id, partialUiDefinition }),
+    enabled: !!partialUiDefinition && !!id,
+  });
+};